src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiofiles==24.1.0
    # via quart
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via quartapp (pyproject.toml)
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
    #   watchfiles
attrs==25.2.0
    # via aiohttp
azure-core==1.32.0
    # via azure-identity
azure-identity==1.21.0
    # via quartapp (pyproject.toml)
blinker==1.9.0
    # via
    #   flask
    #   quart
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   flask
    #   quart
    #   uvicorn
cryptography==44.0.2
    # via
    #   azure-identity
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
flask==3.1.1
    # via quart
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gunicorn==23.0.0
    # via quartapp (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.3.0
    # via hypercorn
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.0
    # via openai
hypercorn==0.17.3
    # via quart
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via
    #   flask
    #   quart
jinja2==3.1.6
    # via
    #   flask
    #   quart
jiter==0.9.0
    # via openai
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   quart
    #   werkzeug
msal==1.32.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
openai==1.108.1
    # via quartapp (pyproject.toml)
packaging==24.2
    # via gunicorn
portalocker==2.10.1
    # via msal-extensions
priority==2.0.0
    # via hypercorn
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
python-dotenv==1.0.1
    # via
    #   quartapp (pyproject.toml)
    #   uvicorn
pyyaml==6.0.2
    # via
    #   quartapp (pyproject.toml)
    #   uvicorn
quart==0.20.0
    # via quartapp (pyproject.toml)
requests==2.32.4
    # via
    #   azure-core
    #   msal
six==1.17.0
    # via azure-core
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-core
    #   azure-identity
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.34.0
    # via quartapp (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
werkzeug==3.1.3
    # via
    #   flask
    #   quart
    #   quartapp (pyproject.toml)
wsproto==1.2.0
    # via hypercorn
yarl==1.18.3
    # via aiohttp
